Lots of stuff. Conceptually, 5 things:
(1) extended lecroy_scope_channel_str() to convert S, T, U, V, W, X, Y, Z
to M1, M2, M3, M4, M5, M6, M7, M8. This allows you to grab traces from the
scope memories. The help text in lgetwf was updated accordingly.

(2) fixed a bug in lecroy_write_wfi_file(), whereby -voffset was being
written, rather than (+)voffset. This bug has always been here, and comes
from the fact that I copied most of the gubbins of this function from
agilent_user.c; Agilent scopes return voffset "upside down." All data
recorded with prior versions of this will have the sign of voffset
recorded incorrectly in the wfi file. (Obviously, if voffset==0, it make
no difference at all).

(3) Added a wrapper to lecroy_write_wfi_file() such that you can force
it to write whatever voffset you like. Backwards compatibility has been
maintained. This extension was necessary because of (4) below.

(4) Added lecroy_subtract_char_arrays(), which subtracts array B from
array A and returns a pointer to the output array. This is used to 
subtract fixed pattern noise (which you record from the scope and put
in array B) from your data (array A). If you save the result (A-B), then
the vertical offset will always be zero (assuming same vertical offset
for both; if the vertical offset isn't the same for both then the result
would be invalid anyway). The function copes with 8-bit or 16-bit data
for all of A, B and the output. Note this function does not talk to the
scope in any way, it just manipulates data. It could be in any library,
I just happen to need it for this one!

(5) Added lecroy_average_segmented_data(). Again, this does not talk to
the scope in any way, so is a generic function that just happens to have
ended up in this library. It takes an input array (of unsigned chars, as
usual) which consists of n traces, generally acquired using segmented
memory. It averages these traces, and returns a pointer to the result
(again of unsigned chars). Can cope with 8-bit or 16-bit data, the
averaged output always has the same number of bytes per point as the
input; so if you want to use averaging to improve your quantisation
noise, always grab the raw data as 16-bit.